requirements.txt
numpy>=1.23
PyYAML>=6.0
matplotlib>=3.7
torch>=2.1
torchinfo>=1.8.0
thop>=0.1.1.post2209072238
hydra-core>=1.3
scipy>=1.10
